Limit the visibility of the DryrunWagedRebalancer so it won't be misused.
